Add first version of 'elistman send'
It still needs work, and I'll eventually move the implementation
somewhere it can be tested. But I have successfully executed this
command against the elistman-dev pipeline running with a DecoyAgent,
getting the response:

  Sent the message successfully to 0 recipients.

So basically, once I switch from a DecoyAgent to a ProdAgent, all the
pieces to actually sign up subscribers and send email will be
operational.