Add `elistman create-subscribers-table`
This replaces bin/create-subscribers-table.sh. It uses db.DynamoDb to
create the table, which sets the schema, the indexes, and the Time To
Live for pending subscribers.

Updated the README to describe how to build the `elistman` CLI and run
it to create the subscribers table. Also updated the description of the
`deploy.env` file.