Have email.SesThrottle manage its own refresh
It became apparent as I was beginning to hook the SesThrottle into
SesMailer that this would be easier. If we want every call before a
SesMailer method to check for a potential refresh, it makes more sense
for it to be built in. Now both the interface and the implementation are
much cleaner and easier to understand and to use.