Use Scan, not DescribeTable, in CountSubscribers
I knew it was too good to be true. The CountSubscribers success case in
dynamodb_contract_test passed when running against a local Docker
instance, but failed against AWS proper. This is due to the well
documented fact that the DescribeTable info is updated only six hours or
so.

Counting via scan still consumes read resources, but it won't send any
item data back:

- https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Scan.html#DDB-Scan-request-Select