Move "Validate SAM template" to build-and-package
Turns out it needs an AWS_REGION configured to run, which I don't want
to do in the run-tests stage of the pipeline. build-and-package seems
the better place to put it.

Also moved the "Build resources" step after the "Assume the testing
pipeline user role" and "Validate SAM template" steps. It seems to make
more sense to run `sam validate` before `sam build`.