Remove EListManStack CloudFormation output
I'd thought I could get filter for EListMan stack instances using this
output, but then realized it would be more complicated than it's worth.

Instead, now that STACK_NAME is part of deploy.env, I'll update the
`elistman send` command to take it as an argument. It'll use a new
function that will lookup the Function output of the stack to get the
Lambda ARN, instead of requiring the raw ARN.