Add EListManFunc interface, NewLambdaFactoryFunc
This is in preparation for injecting an EListManFactoryFunc into
cmd.sendMessage. That will break the dependency on injecting the
CloudFormation and Lambda clients and calling NewLambda directly.
That means the cmd.sendMessage tests, and upcoming `elistman import`
tests, can depend on an even less complex test double, versus
TestLambdaClient and TestCloudFormationClient.