Extract parseSesEvent from newSesEventHandler
This is the first step in a refactoring to eliminate the
baseSesEventHandler subclasses.

I realized while creating the new cliHandler that one cliHandler can
handle multiple event types. It then dawned on me that the current
design of the snsHandler is too complex. It can be simplified by
having one sesHandler with multiple handler functions.