Add SesEventRecord to baseSesEventHandler
It seems logical to just attach the event parsed during parseSesEvent to
the baseSesEventHandler rather than passing it back into its functions.

This required extensive updates to the tests. For one, there's a JSON
string constant or generator function for each type of SesEventRecord
now. newSesEventHandlerFixture takes one of these strings as an argument
to call parseSesEvent.

Though the test changes were more extensive than I expected, the result
is a simplification of how test objects and cases are set up. The
resulting baseSesEventHandler objects are now guaranteed to be
consistent with the events from which they're parsed. This wasn't the
case before, where the test cases had to explicitly assign handler.Type
and so forth explicitly.