Detect goenv shims in bin/check-tools.sh
Per the comment added in this change:

goenv will install a shim into ${GOENV_ROOT}/shims for every tool
installed into ${GOPATH}/bin via `go install`. If the current Go version
doesn't have a tool, but another version does, the shim emits a "command
not found" message to stderr. So we check for "command not found" to
ensure the script will reinstall the tool for the current Go version.