Switch from API Gateway HTTP API to REST API
This is a necessary prerequisite for setting up the AWS Web Application
Firewall (WAF), which only works with REST APIs.

In the end, this wasn't that bad of a change. It required updating the
type name, and adjusting a few fields, but was pretty mechanical and
painless.

The hardest part was figuring out how to give the REST API permission to
write to CloudWatch logs, documented in commit
c63e926f4b6ba042e462f19306c4302c6d035c43.

As a nice bonus, running SAM locally now returns 403 Forbidden for not
found URLs, consistent with prod. Updated bin/smoke-test.sh as a result.