Define a WebACL and association with REST API
This restricts access to the /subscribe endpoint to requests containing
a cookie for a valid CAPTCHA token. (I think—I've yet to see if it works
when it's really live, as noted below.) The rest of the work is (should
be...) on the frontend, which I'll document in the README once I've got
it working to my satisfaction.

I have a local prototype successfully loading the CAPTCHA for the
elistman-dev stack. However, the challenge always fails locally because
it encodes "localhost" as the cookie domain, and the WebACL expects the
live domain. So I need to deploy it to see if it's working completely.
(That's a little trickier than it sounds, because I need to update how I
set my Content-Security-Policy headers to accomodate the new scripts.)

I'll also have to figure out what to do with the bin/smoke-test.sh cases
hitting /subscribe. Most likely I'll update them to confirm that they
fail on the basis of lacking a CAPTCHA token, versus failing due to
being malformed.