Extract ProdAgent.{newTemplate,sendOneEmail}
This is in preparation for updating the `elistman send` command to take
a list of specific addresses as arguments. The first use case will be to
send a test email to oneself before sending it to the whole list.
Another possible use case would be resending an email to folks who
missed it earlier for whatever reason.