Reword verification email to please ignore first
Not sure if this will make much of a difference to avoid complaints from
real recipients who are signed up by a spam bot. But it stands to reason
that if the mail agent shows "Please ignore this email" first, maybe it
will. Hopefully it buys a little trust that the spam was not intentional
from the sender.

Of course, since implementing the CAPTCHA, my spam bot signups have
dropped to zero, but still. Redundant safeguards.