Fold Agent.SendTargeted into Agent.Send
I decided I didn't like the decision whether to send to the entire list
or specific addresses to be made by the the CLI handler. Now there's
just one method that takes the potentially empty list of recipient
addresses. Yay, tests!

I do want to refactor the tests, though, to exercise ProdAgent.send and
ProdAgent.sendTargeted directly. But that's for the next commit.