Consolidate ProdAgent tests, some implementation
The duplication in the tests started to bother me. This is a relatively
unusual case whereby testing several closely intertwined functions
through a single public function is less complicated than testing each
separately. This is because each function is fairly straightforward and
there's a specific dependency pattern between them:

                                 Send
                            (uncomplicated)
                                   |
            Validate message and create message template
                            (uncomplicated)
                                   |
            +------------------------------------------+
            |                                          |
      sendToEntireList                      sendToSpecificRecipients
       (complicated)                             (complicated)
            |                                          |
            +------------------------------------------+
                                   |
                             sendOneEmail
                            (uncomplicated)

Most of the functions are relatively straightforward. The complicated
functions share context (validating the message and creating a template)
and infrastructure (the database and mailer). Both also ultimately call
the same lower level function. The dependencies don't fan out very far,
and the logic fans back into the same fundamental operation.

Consequently, it's less work and less redundant in this case to test
everything through the single Send entrypoint after all. I'm sure
there's an even better, more clear way to express this, but this is the
best I have for now.